Preventing duplicates in persistence.

+++ functions.php
@@ -2,6 +2,8 @@
 	
 	function create_persistence($file='persist.xml', $xml, $persist = 1 )
 	{
+		$exists = false;
+		$ids = array();
 		// Create the file
 		file_put_contents($file, "<?xml version='1.0' encoding='UTF-8'?>\n<persistent_record>\n");
 		// Now we should get the timestamp for this original creation
@@ -11,8 +13,18 @@
 		{
 			// Select a random photo from the possible photos
 			$id = $xml->photos->photo[rand(0, $xml->photos->photo->count()-1)]['id'];
+			foreach($ids as $item) {
+				if($item == $id) {
+					$exists = true; 
+					break;
+				}
+			}
 			// Write the id to the file
-			file_put_contents($file, "<photo id='".$id."'/>\n", FILE_APPEND);
+			if (!$exists) {
+				file_put_contents($file, "<photo id='".$id."'/>\n", FILE_APPEND);
+			}
+			$ids[] = $id;
+			$exists = false;
 		}
 		
 		// Close the root of the XML document
